+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.5] - 2025-09-27
+
+### Fixed
+
+- Custom Entities: Fixed a bug with integer variables
+
+### Added
+
+- Added a check if the source, command or command sequence parameter is empty and return a bad request error
+
+### Changed
+
+- Updated pyyaml to 6.0.3
+
 ## [0.8.4] - 2025-09-24
 
 ### Added